backend/requirements.txt
Django==3.2.18
django-filter==23.1
djangorestframework==3.14.0
djoser==2.2.0
gunicorn==20.1.0
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dotenv
webcolors